Big 5 Inventory

The BFI-2 is a measure of the Big Five personality domains (which we label Extraversion, Agreeableness, Conscientiousness, Negative Emotionality, and Open-Mindedness) and 15 more-specific facet traits. The Big Five personality traits was the model to comprehend the relationship between personality and academic behaviors. This model was defined by several independent sets of researchers who used factor analysis of verbal descriptors of human behavior. These researchers began by studying relationships between a large number of verbal descriptors related to personality traits. They reduced the lists of these descriptors by 5–10 fold and then used factor analysis to group the remaining traits (using data mostly based upon people's estimations, in self-report questionnaire and peer ratings) in order to find the underlying factors of personalityItem numbers for the BFI-2 domain and facet scales are listed below. Reverse-keyed items are denoted by “R.” For more information about the BFI-2, visit the Colby Personality Lab website (http://www.colby.edu/psych/personality-lab/).

Domain Scales

Extraversion: 1, 6, 11R, 16R, 21, 26R, 31R, 36R, 41, 46, 51R, 56 Agreeableness: 2, 7, 12R, 17R, 22R, 27, 32, 37R, 42R, 47R, 52, 57 Conscientiousness: 3R, 8R, 13, 18, 23R, 28R, 33, 38, 43, 48R, 53, 58R Negative Emotionality: 4R, 9R, 14, 19, 24R, 29R, 34, 39, 44R, 49R, 54, 59 Open-Mindedness: 5R, 10, 15, 20, 25R, 30R, 35, 40, 45R, 50R, 55R, 60

Facet Scales

Sociability: 1, 16R, 31R, 46 Assertiveness: 6, 21, 36R, 51R Energy Level: 11R, 26R, 41, 56 Compassion: 2, 17R, 32, 47R Respectfulness: 7, 22R, 37R, 52 Trust: 12R, 27, 42R, 57 Organization: 3R, 18, 33, 48R Productiveness: 8R, 23R, 38, 53 Responsibility: 13, 28R, 43, 58R Anxiety: 4R, 19, 34, 49R Depression: 9R, 24R, 39, 54 Emotional Volatility: 14, 29R, 44R, 59 Intellectual Curiosity: 10, 25R, 40, 55R Aesthetic Sensitivity: 5R, 20, 35, 50R Creative Imagination: 15, 30R, 45R, 60

Relational table

Domain	Factor-pure facet	Complementary facets

E	Sociability	Assertiveness, Energy Level

A	Compassion	Respectfulness, Trust

C	Organization	Productiveness, Responsibility

N	Anxiety	Depression, Emotional Volatility

O	Aesthetic Sensitivity	Intellectual Curiosity, Creative Imagination

Data requirements

Column names

The package functions expect the data to be named in a specific way, and to not contain data other than the BFI-2 data. Column names should be zero-leading two digits to indicate the question number, and they should end with these two digits. If this system is followed, then all functions work out of the box.

Examples that work:

bfi_01 bfi_02 ... bfi_59 bfi_60

big_five_01 big_five_02 ... big_five_59 bbig_five_60

Examples that won't work

bfi_1 bfi_2 ... bfi_59 bfi_60

big_five_01_trust big_five_02_change ... big_five_59_test bbig_five_60_lat

Data values

The data should be coded with the original scoring system 1-5. The data should not have implemented necessary reversal of answers for any of the questions, the functions will take care of this.

References

Soto, C. J., & John, O. P. (2017). The next Big Five Inventory (BFI-2): Developing and assessing a hierarchical model with 15 facets to enhance bandwidth, fidelity, and predictive power. Journal of Personality and Social Psychology, 113(1), 117–143. https://doi.org/10.1037/pspp0000096

bfi_compute(
  data,
  type = c("domains", "facets"),
  keep_all = FALSE,
  prefix = "bfi_"
)

bfi_compute_domains(
  data,
  domains = c("extraversion", "agreeableness", "conscientiousness",
    "negative emotionality", "open-mindedness"),
  keep_all = FALSE,
  prefix = "domain_"
)

bfi_compute_facets(
  data,
  facets = c("sociability", "assertiveness", "energy", "compassion", "respectful",
    "trust", "organization", "productive", "responsibility", "anxiety", "depression",
    "emotional volatility", "intellectual curiosity", "aesthetic sensebility",
    "creative imagination"),
  keep_all = FALSE,
  prefix = "facet_"
)

Arguments

data

data.frame containing bfi data

type

Choose domains or facets. Default is both

keep_all

logical, append to data.frame

prefix

string to prefix column names of computed values

domains

string vector of domains to compute

facets

string vector of facets to compute

Value

data.frame with calculated scores

Functions

bfi_compute_domains(): Compute BFI-2 domains and return in a data.frame

bfi_compute_facets(): Compute BFI-2 domains and return in a data.frame

Examples

library(dplyr)
# Making some test data
test_data <- tibble(
  id = rep(1:10, each = 60),
  name = rep(sprintf("bfi_%02d", 1:60), 10),
  value = lapply(1:10, function(x){
    sample(1:5, size = 60, replace = TRUE)
  }) %>% unlist()
) %>% 
  tidyr::pivot_wider()
  
bfi_compute(test_data)
#>    bfi_domain_extrav bfi_domain_agree bfi_domain_consc bfi_domain_negem
#> 1                 27               38               41               38
#> 2                 36               30               46               30
#> 3                 34               41               39               41
#> 4                 32               46               39               46
#> 5                 34               27               36               27
#> 6                 28               50               44               50
#> 7                 32               38               44               38
#> 8                 31               36               24               36
#> 9                 31               33               37               33
#> 10                36               37               34               37
#>    bfi_domain_openm bfi_facet_sociabi bfi_facet_asserti bfi_facet_energy
#> 1                32                 9                 9                9
#> 2                36                16                11                9
#> 3                36                15                 9               10
#> 4                38                12                12                8
#> 5                31                10                12               12
#> 6                35                 8                10               10
#> 7                47                11                12                9
#> 8                47                10                10               11
#> 9                46                12                12                7
#> 10               38                11                11               14
#>    bfi_facet_compass bfi_facet_respect bfi_facet_trust bfi_facet_organiz
#> 1                 16                15               7                12
#> 2                 13                10               7                14
#> 3                 14                15              12                11
#> 4                 16                15              15                12
#> 5                  6                10              11                 9
#> 6                 15                17              18                13
#> 7                 10                12              16                16
#> 8                 11                14              11                 5
#> 9                 15                 9               9                11
#> 10                11                12              14                14
#>    bfi_facet_product bfi_facet_respons bfi_facet_anxiety bfi_facet_depress
#> 1                 13                16                 9                12
#> 2                 20                12                12                13
#> 3                 11                17                12                15
#> 4                 15                12                11                11
#> 5                  9                18                16                 7
#> 6                 16                15                16                 9
#> 7                 12                16                16                15
#> 8                 11                 8                16                18
#> 9                 14                12                14                15
#> 10                10                10                15                11
#>    bfi_facet_emovola bfi_facet_intcuri bfi_facet_aessens bfi_facet_cretima
#> 1                 11                11                 6                11
#> 2                 11                14                13                13
#> 3                  9                14                16                14
#> 4                 16                13                12                12
#> 5                  8                12                15                13
#> 6                 10                 9                10                11
#> 7                 16                18                15                15
#> 8                 13                 9                13                 7
#> 9                 17                14                 7                15
#> 10                12                10                14                15
bfi_compute(test_data, prefix = "bfi_")
#>    bfi_domain_extrav bfi_domain_agree bfi_domain_consc bfi_domain_negem
#> 1                 27               38               41               38
#> 2                 36               30               46               30
#> 3                 34               41               39               41
#> 4                 32               46               39               46
#> 5                 34               27               36               27
#> 6                 28               50               44               50
#> 7                 32               38               44               38
#> 8                 31               36               24               36
#> 9                 31               33               37               33
#> 10                36               37               34               37
#>    bfi_domain_openm bfi_facet_sociabi bfi_facet_asserti bfi_facet_energy
#> 1                32                 9                 9                9
#> 2                36                16                11                9
#> 3                36                15                 9               10
#> 4                38                12                12                8
#> 5                31                10                12               12
#> 6                35                 8                10               10
#> 7                47                11                12                9
#> 8                47                10                10               11
#> 9                46                12                12                7
#> 10               38                11                11               14
#>    bfi_facet_compass bfi_facet_respect bfi_facet_trust bfi_facet_organiz
#> 1                 16                15               7                12
#> 2                 13                10               7                14
#> 3                 14                15              12                11
#> 4                 16                15              15                12
#> 5                  6                10              11                 9
#> 6                 15                17              18                13
#> 7                 10                12              16                16
#> 8                 11                14              11                 5
#> 9                 15                 9               9                11
#> 10                11                12              14                14
#>    bfi_facet_product bfi_facet_respons bfi_facet_anxiety bfi_facet_depress
#> 1                 13                16                 9                12
#> 2                 20                12                12                13
#> 3                 11                17                12                15
#> 4                 15                12                11                11
#> 5                  9                18                16                 7
#> 6                 16                15                16                 9
#> 7                 12                16                16                15
#> 8                 11                 8                16                18
#> 9                 14                12                14                15
#> 10                10                10                15                11
#>    bfi_facet_emovola bfi_facet_intcuri bfi_facet_aessens bfi_facet_cretima
#> 1                 11                11                 6                11
#> 2                 11                14                13                13
#> 3                  9                14                16                14
#> 4                 16                13                12                12
#> 5                  8                12                15                13
#> 6                 10                 9                10                11
#> 7                 16                18                15                15
#> 8                 13                 9                13                 7
#> 9                 17                14                 7                15
#> 10                12                10                14                15