Edinburgh handedness inventory

Compute all variables of ehi, using other functions in this package. Will return the given data.frame with three additional columns, the laterality quotient (LQ), the laterality factor (Coded), and the nominal laterality code (Nominal).

ehi_compute(
  data,
  cols = matches("^ehi_[0-9][0-9]$"),
  writing = ehi_01,
  ...,
  keep_all = TRUE,
  prefix = "ehi_"
)

Arguments

data

data.frame containing ehi data

cols

tidyselected columns of all ehi data

writing

numeric vector of writing preference (-2,-1,0,1,2)

...

additional arguments to ehi_factorise_lqa

keep_all

logical, append to data.frame

prefix

string to prefix column names of computed values

Value

data.frame

Details

Background

The Edinburgh Handedness Inventory is a measurement scale used to assess the dominance of a person's right or left hand in everyday activities, sometimes referred to as laterality. The inventory can be used by an observer assessing the person, or by a person self-reporting hand use. The latter method tends to be less reliable due to a person over-attributing tasks to the dominant hand.

Scoring

The EHI has several measures that can help assess a person's laterality.

answer	value	nominal	lq	lq_cat	lqa_cat

Left dominance	-2	left	-100	left	left

Left preference	-1	left	-40	left	ambidexter

No preference	0	ambidexter	0	right	ambidexter

Right preference	1	right	40	right	ambidexter

Right dominance	2	right	100	right	right

Nominal

The easiest measure from the EHI is the nominal laterality value, which is just the answer to the first question on hand preference when writing. This simple index just treat negative answers as "left" dominance, positive number as "right" dominance, and a 0 as ambidextrous. Note: The original paper by Oldfield (1971) does not explicitly state a category for "Ambidextrous". It is very rare that a person does not have a clear preference on writing hand, even if they can write with both hands. This category is only added in this package to handle the possible case of someone answering "No preference".

min	max	category

-2	-1	left

0	0	ambidexter

1	2	right

Laterality quotient (lq)

The total score of the EHI is more than just summing the values for each answer. The laterality quotient (LQ) uses the answers to all the questions. The LQ can take values from -100 to 100, and is calculated by taking the sum of all positive answers subtracting the sum of absolute values of the negative answers, divided by the sum of both, and multiplied by 100.

katex::katex_html(equation)

Σ(∣positive∣)−Σ(∣negative∣)Σ(∣positive∣)+Σ(∣negative∣)x100\frac{\Sigma(|positive|)-\Sigma(|negative|)}{\Sigma(|positive|)+\Sigma(|negative|)}x100Σ(∣positive∣)+Σ(∣negative∣)Σ(∣positive∣)−Σ(∣negative∣)​x100

Laterality index

The laterality index is based on the laterality quotient (above) and categorises answers into to categories, Left and Right. The Oldfield (1971) paper mentions "indeterminate handedness" a couple of times in the paper, but the case for "true" ambidextrous is not made, and as such the inventory does not have official categories for that. As the index is based on the quotient, that ranges from -100 to 100, getting a perfect 0 LQ is very unlikely, and as indicated in the paper, such score is assumed to belong to the Right hand part of the scale.

min	max	category

-100	-1	left

0	100	right

An alternate laterality index is also often employed, where scores between -40 and 40 are treated as ambidextrous.

Data requirements

One row of data should refer to a single questionnaire answered, and as such, if a person has answered multiple times, these should appear on separate rows with columns identifying ID and time point per observation.

Column names

For ease, we recommend naming the columns in a consistent way, so the functions in this package become easier to use. The LCBC database follows a naming scheme that prefixes all columns with ehi_ and ends with a zero-padded double digit indicator of the question number.

Data values

The cell values in the data should be coded from -2 through 0 to 2, and there should be a single value per question.

value	category

-2	Left hand dominance

-1	Left hand preference

0	No preference

1	Right hand preference

2	Right hand dominance

References

Oldfield, RC (March 1971) The assessment and analysis of handedness: The Edinburgh inventory. Neuropsychologia. 9 (1): 97–113. doi:10.1016/0028-3932(71)90067-4

Verdino, M; Dingman, S (April 1998). Two measures of laterality in handedness: the Edinburgh Handedness Inventory and the Purdue Pegboard test of manual dexterity. Perceptual and Motor Skills. 86 (2): 476–8. doi:10.2466/pms.1998.86.2.476

Knecht, S; Dräger, B; Deppe, M; Bobe, L; Lohmann, H; Flöel, A; Ringelstein, E-B; Henningsen, H (December 2000). Handedness and hemispheric language dominance in healthy humans. Brain. 123 (12): 2512–8. doi:10.1093/brain/123.12.2512.

See also