Compute met from IPAQ

The purpose of the International Physical Activity questionnaires (IPAQ) is to provide a set of well-developed instruments that can be used internationally to obtain comparable estimates of physical activity. There are two versions of the questionnaire. The short version is suitable for use in national and regional surveillance systems and the long version provide more detailed information often required in research work or for evaluation purposes.

Scoring

Scoring of the IPAQ is based on a metric called METs, which are multiples of the resting metabolic rate. The IPAQ scoring description can be found here

Continuous Score

Expressed as MET-min per week: MET level x minutes of activity/day x days per week

MET levels:

Light - 3.3 METs

Moderate - 4.0 METs

Vigorous - 8.0 METs

Total MET-minutes/week = Light (3.3 x min x days) + Mod (4.0 x min x days) + Vig (8.0 x min x days)

Categorical Score

Three levels (categories) of physical activity are proposed:

Category 1: Low

This is the lowest level of physical activity. Those individuals who not meet criteria for categories 2 or 3 are considered low/inactive.

Category 2: Moderate

Any one of the following 3 criteria:

3 or more days of vigorous activity of at least 20 minutes per day OR

5 or more days of moderate-intensity activity or walking of at least 30 minutes per day OR

5 or more days of any combination of walking, moderate-intensity or vigorous intensity activities achieving a minimum of at least 600 MET-min/week.

Category 3: High

Any one of the following 2 criteria:

Vigorous-intensity activity on at least 3 days and accumulating at least 1500 MET-minutes/ week OR

7 or more days of any combination of walking, moderate-intensity or vigorous intensity activities achieving a minimum of at least 3000 MET-minutes/week

References

Depression Screening Scale: A Preliminary Report, J Psychiatr Res, 17 (1), 37-49, doi: 10.1016/0022-3956(82)90033-4

E L Lesher 1, J S Berryhill (1994), Validation of the Geriatric Depression Scale -- Short Form Among Inpatients, J Clin Psychol, 50 (2), 256-60, doi: 10.1002/1097-4679(199403)50:2<256::aid-jclp2270500218>3.0.co;2-e

ipaq_compute_met(minutes = ipaq_2, days = ipaq_1b, met = ipaq_mets()$light)

ipaq_compute_sum(vigorous, moderate, light)

ipaq_compute(
  data,
  mets = ipaq_mets(),
  light_days = ipaq_5b,
  light_mins = ipaq_6,
  mod_days = ipaq_3b,
  mod_mins = ipaq_4,
  vig_days = ipaq_1b,
  vig_mins = ipaq_2,
  prefix = "ipaq_",
  keep_all = TRUE
)

Arguments

minutes

vector of numeric minutes

days

vector of numeric days

met

met number (light = 3.3, moderate = 4.0, vigorous = 8)

vigorous

Vector with vigorous met calculated

moderate

Vector with moderate met calculated

light

Vector with light met calculated

data

data.frame containing all the ipaq data

mets

list generated with ipaq_mets() (default = ipaq_mets())

light_days

column with the days of light activity

light_mins

column with the minutes of light activity

mod_days

column with the days of moderate activity

mod_mins

column with the minutes of moderate activity

vig_days

column with the days of vigorous activity

vig_mins

column with the minutes of vigorous activity

prefix

string to prefix column names of computed values

keep_all

logical, append to data.frame

Value

data.frame

Functions

ipaq_compute_met(): Calculate mets of an activity type

ipaq_compute_sum(): Calculate the IPAQ sum based on activities and mets

See also

Examples

ipaq_vig_mins <- c(60, 20, 60, 25, 90, 20, 0, 75, 60, 30)
ipaq_vig_days <- c(1, 3, 2, 5, 6, 1, 1, 2, 2, 4)
ipaq_compute_met(ipaq_vig_mins, ipaq_vig_days, met = 8.0)
#>  [1]  480  480  960 1000 4320  160    0 1200  960  960
light = c(1300, 300)
moderate = c(200, 400)
vigorous = c(0, 1300)
ipaq_compute_sum(vigorous , moderate, light)
#> [1] 1500 2000