Zygocity - Calculate item

Classification

This note contains a brief description of the algorithm used to determine zygocity in recruitment in the 2000s.

Name	Answer questions about...	Used for

Drop	You and your twin were like two drops of water in childhood	Pairs and singles

Stranger	Strangers had trouble telling the difference when you were children	Pairs and singles

Eye	Similarity in terms of eye color	Pairs

Voice	Similarity in terms of voice	Single

Dexter	Similarity in Dexterity	Pairs and Singles

Belief	What you believe yourself	Pairs and Singles

"Single" twins here means those who have responded alone, i.e. there is no data available for both in the pair. The similarity questions that are not found in the table above, e.g. whether or not family members had problems distinguishing the twins is not used in the classification.

Weights

During calculations of the entire zygocity score, weights are applied to the different categories, depending on whether one or both twins have responded to the questionnaire.

Name	Answer questions about...	Factor single	Factor pair

Drop	You and your twin were like two drops of water	1.494	2.111

Stranger	Strangers had trouble seeing the difference	0.647	0.691

Eye	Similarity in terms of eye color		0.394

Voice	Similarity in terms of voice	0.347

Dexter	Dexterity Similarity	0.458	0.366

Belief	What you believe yourself	0.417	0.481

Constant term in the formula	0.007	- 0.087

zygo_calc(x, type, n = "single", recode = TRUE)

Arguments

x

integer vector of answers to one of the questionnaire questions. Should not be longer than 2.

type

type of question the vector is from. "drop", "stranger, "dexterity", "voice", "eye", or "belief".

n

string indicating number of twins in the pair available. Either "single" or "pair".

recode

logical indicating if data should be recoded from 1-5(7) to -1. 0. 1.

Value

single value of calculated score based on recoded vector and multiplied with correct factor weight.

Examples

zygo_calc(c(1), type = "eye")
#> Warning: Cannot calculate 'eye' for 'single' data
#> [1] NA
zygo_calc(c(1,3), type = "belief", n = "pair")
#> [1] 0.2405
zygo_calc(c(4), type = "voice")
#> [1] 0